# crates/twerk-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Twerk API",
    "description": "Task scheduling and job queue management API",
    "contact": {
      "name": "Twerk Contributors"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "tags": [
          "super::handlers::jobs::read"
        ],
        "summary": "GET /jobs",
        "operationId": "list_jobs_handler",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JobSummary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "super::handlers::jobs::create"
        ],
        "summary": "POST /jobs",
        "operationId": "create_job_handler",
        "parameters": [
          {
            "name": "wait",
            "in": "query",
            "description": "Whether to block until the job completes (true/false/blocking)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Job definition as JSON or YAML",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            },
            "application/x-yaml": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            },
            "application/yaml": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            },
            "text/yaml": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Job created or completed when wait=blocking",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid job definition or unsupported content type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{id}/cancel": {
      "put": {
        "tags": [
          "super::handlers::jobs::mutation"
        ],
        "summary": "PUT /jobs/{id}/cancel",
        "operationId": "cancel_job_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The job ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/JobId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Job cannot be cancelled in its current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "super::handlers::jobs::mutation"
        ],
        "summary": "POST /jobs/{id}/cancel",
        "operationId": "cancel_job_handler_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The job ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/JobId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Job cannot be cancelled in its current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
...
    "/jobs/{id}/restart": {
      "put": {
        "tags": [
          "super::handlers::jobs::mutation"
        ],
        "summary": "PUT /jobs/{id}/restart",
        "operationId": "restart_job_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The job ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/JobId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job restarted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Job cannot be restarted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutoDelete": {
        "type": "object",
        "description": "`AutoDelete` defines automatic cleanup configuration.",
        "properties": {
          "after": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "CreateJobResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/JobSummary"
          },
          {
            "$ref": "#/components/schemas/Job"
          }
        ]
      },
...
      "Job": {
        "type": "object",
        "description": "Job represents a job in the system.",
        "properties": {
          "autoDelete": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AutoDelete"
              }
            ]
          },
          "completedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "context": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobContext"
              }
            ]
          },
          "createdAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdBy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "defaults": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobDefaults"
              }
            ]
          },
          "deleteAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "execution": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "failedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobId"
              }
            ]
          },
          "inputs": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "output": {
            "type": [
              "string",
              "null"
            ]
          },
          "parentId": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobId"
              }
            ]
          },
          "permissions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "position": {
            "type": "integer",
            "format": "int64"
          },
          "progress": {
            "type": "number",
            "format": "double"
          },
          "result": {
            "type": [
              "string",
              "null"
            ]
          },
          "schedule": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobSchedule"
              }
            ]
          },
          "secrets": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "startedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "taskCount": {
            "type": "integer",
            "format": "int64"
          },
          "tasks": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "webhooks": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Webhook"
            }
          }
        }
      },
      "JobContext": {
        "type": "object",
        "description": "`JobContext` holds contextual information for a job.",
        "properties": {
          "inputs": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "job": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "secrets": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "tasks": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
      "JobDefaults": {
        "type": "object",
        "description": "`JobDefaults` defines default values for job tasks.",
        "properties": {
          "limits": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TaskLimits"
              }
            ]
          },
          "priority": {
            "type": "integer",
            "format": "int64"
          },
          "queue": {
            "type": [
              "string",
              "null"
            ]
          },
          "retry": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TaskRetry"
              }
            ]
          },
          "timeout": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "JobId": {
        "type": "string",
        "description": "RFC 4122 UUID or 22-character base57 short job ID emitted by the coordinator",
        "pattern": "^(?:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[23456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz]{22})$"
      },
...
      "JobSchedule": {
        "type": "object",
        "description": "`JobSchedule` defines a job schedule.",
        "properties": {
          "cron": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/string"
              }
            ]
          }
        }
      },
      "JobState": {
        "type": "string",
        "description": "State machine for a [`Job`].\n\nValid transitions:\n- `Pending` -> `Scheduled`\n- `Scheduled` -> `Running`\n- `Running` -> `Completed` | `Failed` | `Cancelled`\n- `Failed` | `Cancelled` -> `Restart`\n- `Restart` -> `Pending`",
        "enum": [
          "PENDING",
          "SCHEDULED",
          "RUNNING",
          "CANCELLED",
          "COMPLETED",
          "FAILED",
          "RESTART"
        ]
      },
...
      "MessageResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "Page_JobSummary": {
        "type": "object",
        "description": "Page represents a paginated result set",
        "required": [
          "items",
          "number",
          "size",
          "total_pages",
          "total_items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "`JobSummary` provides a summary view of a job.",
              "properties": {
                "completedAt": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "createdAt": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "createdBy": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/User"
                    }
                  ]
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "error": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "failedAt": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "id": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/JobId"
                    }
                  ]
                },
                "inputs": {
                  "type": [
                    "object",
                    "null"
                  ],
                  "additionalProperties": {
                    "type": "string"
                  },
                  "propertyNames": {
                    "type": "string"
                  }
                },
                "name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "parentId": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/JobId"
                    }
                  ]
                },
                "position": {
                  "type": "integer",
                  "format": "int64"
                },
                "progress": {
                  "type": "number",
                  "format": "double"
                },
                "result": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "schedule": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/JobSchedule"
                    }
                  ]
                },
                "startedAt": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "state": {
                  "$ref": "#/components/schemas/JobState"
                },
                "tags": {
                  "type": [
                    "array",
                    "null"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "taskCount": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            },
            "description": "Items in the current page"
          },
          "number": {
            "type": "integer",
            "format": "int64",
            "description": "Current page number"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "Page size"
          },
          "total_items": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of items"
          },
          "total_pages": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of pages"
          }
        }
      },
...
      "Permission": {
        "type": "object",
        "description": "Permission defines access permissions.",
        "properties": {
          "role": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "user": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          }
        }
      },
...
      "Role": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "slug": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "StatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "TaskLimits": {
        "type": "object",
        "description": "`TaskLimits` defines resource limits for a task.",
        "properties": {
          "cpus": {
            "type": [
              "string",
              "null"
            ]
          },
          "memory": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TaskRetry": {
        "type": "object",
        "description": "`TaskRetry` defines retry configuration for a task.",
        "properties": {
          "attempts": {
            "type": "integer",
            "format": "int64"
          },
          "limit": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "disabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Webhook": {
        "type": "object",
        "description": "Webhook defines webhook notification configuration.",
        "properties": {
          "event": {
            "type": [
              "string",
              "null"
            ]
          },
          "headers": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "if": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "string": {
        "type": "string"
      }
    }
  }
}